requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
certifi==2018.8.24
chardet==3.0.4
configparser==3.5.0
enum34==1.1.6
futures==3.2.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.9.3
python-dotenv==0.9.1
requests==2.19.1
singledispatch==3.4.0.3
six==1.11.0
slackclient==1.2.1
urllib3==1.23
websocket-client==0.53.0
wrapt==1.10.11
